Move ECDSA functionality into ECDSA module
